Add Unit Testing for the ProcessCommandLine function.
We add test code for the command line arguments passed to
our cipher. Tow versions are included, to better grasp the possible
cases of cmd args. passed.
The CMakeLists for the test files now includes a definition for
an 'interface' library linked. This simplifies the CMakeLists file, when multiple
tests are included.